Fix intra-doc links
When deployed to GitHub pages, some of the links were broken, due to a
missing ../ in the url. I thought Docusaurus would check for this, but
apparently it did not. Change "warn" to "throw", maybe that would have
detected it. And change intra-doc links to have a .md suffix. Previously
this broke something for me, but now it seems to work fine.